backend/requirements.txt
fastapi
uvicorn
sqlmodel
httpx
jinja2
pydantic
python-multipart
sqlalchemy
requests
python-dotenv
alembic
openai
huggingface_hub